Usuario no se puede eliminar así mismo y arreglado problema con eliminar partida

+++ src/main/java/org/springframework/samples/petclinic/jugador/JugadorController.java
@@ -215,7 +215,7 @@ public class JugadorController {
 
 	
 
-	//HAY QUE CONTROLAR SI TIENE ESE USUARIO UNA PARTIDA JUGADA, PODER ELIMINARLO TAMBIÉN
+	
 	@GetMapping(path = "/jugador/delete/{id}")
 	public ModelAndView deleteGame(@PathVariable("id") int id, ModelMap modelMap) {
 	Authentication auth = SecurityContextHolder.getContext().getAuthentication();
@@ -226,10 +226,17 @@ public class JugadorController {
 			String credencial = usuarioR.getAuthority();
 			if (credencial.equals("admin")) {
 				Jugador jugador = jugadorService.findJugadorById(id);
-				jugadorService.deleteJugador(jugador);
-				ModelAndView result = new ModelAndView("users/UsersList");
-				result.addObject("users", authoritiesService.findAllUsers());
-				return result;
+				if(jugador.getUser().getUsername() == currentUser.getUsername()){
+					ModelAndView result = new ModelAndView("users/UsersList");
+					result.addObject("users", authoritiesService.findAllUsers());
+					return result;
+				} else {
+					jugadorService.deleteJugador(jugador);
+					ModelAndView result = new ModelAndView("users/UsersList");
+					result.addObject("users", authoritiesService.findAllUsers());
+					return result;
+				}
+				
 			}
 		}
